Rename Init to Initialize; add isNew parameter
Renamed the `Init` method to `Initialize` in the `IBlogConfigInitializer` interface and `BlogConfigInitializer` class. The method now accepts a `bool` parameter named `isNew`.

Updated `BlogConfigInitializer.Initialize` to conditionally use `SystemManifestSettings.DefaultValueNew` if `isNew` is `true`, otherwise it uses `SystemManifestSettings.DefaultValue`.

Added a new static property `DefaultValueNew` to the `SystemManifestSettings` class, initializing `VersionString` to `"0.0.0.0"` and `InstallTimeUtc` to the current UTC time.