Remove class changes for images in dark/light mode
The `useDarkMode` and `useLightMode` functions in `theme.module.js`
were updated to stop adding/removing `bg-dark` and `border-secondary`
classes to/from images with the `img-thumbnail` class inside elements
with the `post-content` class. This change ensures that these specific
class modifications for images do not occur when toggling between dark
and light modes.