Refactor resizeImages to use vanilla JavaScript
Replaced jQuery methods with native JavaScript in the `resizeImages` function. This change removes the dependency on jQuery, improving performance and reducing the need for an external library. Specifically, `document.querySelectorAll` is used to select image elements within `.post-content`, and `forEach` is used to iterate over these elements to remove their `height` and `width` attributes and add the `img-fluid` and `img-thumbnail` classes.