Add experimental feature to use Gravatar as avatar
Introduce a new experimental feature controlled by the
`Experimental:UseOwnerEmailAsAvatar` setting in `appsettings.json`.
Inject `IConfiguration` in `_WidgetsProfile.cshtml` to access the
configuration. Define `useOwnerEmailAsAvatar` to read the setting's
value. Add conditional logic to display either a Gravatar image based
on the owner's email or the existing avatar URL. Use the `gravatar` tag
when `useOwnerEmailAsAvatar` is true and the `img` tag when false.