Update SQL Server migration script URL in appsettings.json
The `DatabaseMigrationScript` URL for `SqlServer` in the `appsettings.json` file has been updated. The URL was changed from pointing to the `master` branch to the `release` branch of the repository. This change ensures that the SQL Server migration script is now fetched from the `release` branch instead of the `master` branch.