Expand ID range for custom themes in GetStyleSheetQueryHandler
The condition in the `if` statement within the `Handle` method of the
`GetStyleSheetQueryHandler` class has been modified. Previously, the
condition checked if `request.Id` was less than 100. The updated
condition now checks if `request.Id` is either less than 100 or greater
than 110. This change likely aims to expand the range of IDs that are
considered custom themes, with IDs outside the range of 100 to 110 being
treated as custom themes.