Prevent migration on non-stable versions; update theme CSS
Added a check in `MigrationManager.cs` to prevent database migration if the application is running a non-stable version, using the new `Helper.IsNonStableVersion()` method. Logged a warning if the version is non-stable.

Introduced `IsNonStableVersion()` in `Helper.cs` to identify non-stable versions using a regular expression.

Updated `Appearance.cshtml` to enhance theme selector styling:
- Changed border color to `--accent-color1`.
- Removed `text-light` class from `theme-color-indicator` divs.
- Applied a linear gradient with `--accent-color1` to the background of `theme-color-indicator` divs.
- Ensured consistent styling in the theme selector footer with `--accent-color1` for the background color.