Add version query string to TinyMCE script in EditPost.cshtml
Updated the script tag for loading the TinyMCE library in the
EditPost.cshtml file by adding the asp-append-version="true"
attribute. This change ensures that a version query string is
appended to the script URL, which helps with cache busting by
forcing the browser to load the latest version of the script
whenever it changes.