Drop 'Revision' column and create 'PostView' table
The code first checks if the column 'Revision' exists in the table '[dbo].[Post]'. If it does, the column is dropped.
The code then checks if the table '[dbo].[PostView]' does not exist. If it does not exist, the code creates the table '[dbo].[PostView]' with the following columns:
  * 'PostId' of type 'uniqueidentifier' which is not nullable.
  * 'RequestCount' of type 'int' which is not nullable.
  * 'ViewCount' of type 'int' which is not nullable.
The table '[dbo].[PostView]' is created with a primary key constraint on the 'PostId' column, and various table options are set for the primary key.