Add PostView property and fetch data in OnGetAsync
The `PostModel` class now includes a `PostView` property of type `PostViewEntity`. In the `OnGetAsync` method, after setting the `Post` property and updating the `ViewData["TitlePrefix"]`, a new line has been added to fetch the `PostView` data using the `mediator` and the `GetPostViewQuery` with the `post.Id` as a parameter. This ensures that the view has all necessary information when the page is rendered.